MYR-806: moved stories to component dirs and fixed build error
