+++ src/components-v2/RightMenuBar/index.ts
@@ -1,0 +1,3 @@
+import {RightMenuBarProps} from './right-menubar.interfaces';
+
+export type {RightMenuBarProps};
