+++ src/components-v2/RightMenuBar/right-menubar.interfaces.ts
@@ -1,0 +1,7 @@
+import {ExperienceDummy} from '../ExperienceList/';
+
+interface RightMenuBarProps {
+  experiences: ExperienceDummy[];
+}
+
+export type {RightMenuBarProps};
